// frontend/src/api/model/patchedOrganizationDetailRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DOSPORTAL API
 * API Documentation for DOSPORTAL
 * OpenAPI spec version: 1.0.0
 */
import type { DataPolicyEnum } from './dataPolicyEnum';

export interface PatchedOrganizationDetailRequest {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name?: string;
  data_policy?: DataPolicyEnum;
  /**
   * @maxLength 200
   * @nullable
   */
  website?: string | null;
  /**
   * @maxLength 200
   * @nullable
   */
  contact_email?: string | null;
  /** @nullable */
  description?: string | null;
}